requirements.txt
crewai>=0.28.0
openai>=1.14.0
supabase>=2.4.0
resend>=0.7.0
notion-client>=2.2.1
apscheduler>=3.10.4
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.1
